Moves make local and clear assets to fork
Moves the make_all_local_and_clear_assets function from the linking service to a new fork service.
This change improves code organization and separates concerns related to data forking and linking.
Removes unused import bpy from linking.
